+++ bogofilter/src/buff.h
@@ -19,8 +19,8 @@ AUTHOR:
 
 typedef struct {
     word_t t;
-    size_t read;	/* start of last read */
-    size_t size;	/* capacity */
+    uint   read;	/* start of last read */
+    uint   size;	/* capacity */
 } buff_t;
 
 extern buff_t  *buff_new(byte *buff, size_t used, size_t size);
